+++ test/hexapdf/type/test_page_tree_node.rb
@@ -217,9 +217,19 @@ describe HexaPDF::Type::PageTreeNode do
       assert(@root.validate)
     end
 
+    it "works if the location stays the same" do
+      @root.move_page(3, 3)
+      assert_equal(@pages, @root.each_page.to_a)
+      assert(@root.validate)
+
+      @root.move_page(-2, -2)
+      assert_equal(@pages, @root.each_page.to_a)
+      assert(@root.validate)
+    end
+
     it "moves the page to the correct location with a positive index" do
       @root.move_page(1, 3)
-      assert_equal([@pages[0], @pages[2], @pages[1], *@pages[3..-1]], @root.each_page.to_a)
+      assert_equal([@pages[0], @pages[2], @pages[3], @pages[1], *@pages[4..-1]], @root.each_page.to_a)
       assert(@root.validate)
     end
 
@@ -231,11 +241,11 @@ describe HexaPDF::Type::PageTreeNode do
 
     it "moves the page to the correct location within the same parent node" do
       @root.move_page(2, 4)
-      assert_equal([@pages[0], @pages[1], @pages[3], @pages[2], *@pages[4..-1]], @root.each_page.to_a)
+      assert_equal([@pages[0], @pages[1], @pages[3], @pages[4], @pages[2], *@pages[5..-1]], @root.each_page.to_a)
       assert(@root.validate)
 
       @root.move_page(4, 3)
-      assert_equal([@pages[0], @pages[1], @pages[3], @pages[4], @pages[2], *@pages[5..-1]], @root.each_page.to_a)
+      assert_equal([@pages[0], @pages[1], @pages[3], @pages[2], @pages[4], *@pages[5..-1]], @root.each_page.to_a)
       assert(@root.validate)
     end